Cleaned up code / Added walls to genblocks
The wall floor and roof are now part of the genblock loop, and a lot of
the extra left over variables we have were deleted. However something
that I think was done by jia made it so the player only spawns at 0,0
causing him to fall through the map. I don't know what is causing this.