THE BLOCKS.... THEY MOVE!
THEY WILL TAKE OVER THE WORLD! PLS SEND HALP! THE BLOCKS WILL MOVE THE
WORLD.

But yeah I fixed the movable block class so when you hold f and move in
the direction of the block, you move with half the speed with the block.
Yu can't phase through the block. The block can't fall (there is no
gravity on it due to reasons) otherwise it still works. I added it to
the Help.txt in the mapeditor (not the official one). I also added a
texture.